don't force a specific applicationCode
It appears that the application code is generally set to whatever application Inspektr is running in.
i.e. CAS or CAS_Management

We shouldn't force our own applicationCode, but rather let the application using the library determine it. But we should still include that the audits came from the IDM management library, which is why I prefixed actions with IDM_